@@ -5,7 +5,7 @@ import typescript from '@rollup/plugin-typescript';
 import dts from 'rollup-plugin-dts';
 import { terser } from 'rollup-plugin-minification';
 
-const packageJson = require('./package.json');
+import packageJson from './package.json' assert { type: 'json' };
 
 export default [
     {
